Split declaration and definition of svd related functions
